refactor(ui): add SearchForm with fixed height for Text widget
Defined a dedicated SearchForm to limit the height of the search form
in the Text widget, which currently supports only one line. The height
of the search form is fixed at 1 since it does not support borders yet.
In the future, when border support is added, the form height will be
adjusted appropriately (1-3) based on the presence of borders.